Change ScrollDown to Accept, & Logout on cancel
The button "Scroll Down" can be clicked all the time and opens a popup
"Scroll down to accept". Should be renamed to "Accept", disabled until
scrolled down and all cancel actions should lead to logout.

After accepting, the popup still appears for a not so short moment when
loading a pege.